+++ src/mantle/database/factory/class-term-factory.php
@@ -51,7 +51,7 @@ class Term_Factory extends Factory {
 	public function definition(): array {
 		return [
 			'description' => trim( $this->faker->paragraph( 2 ) ),
-			'name'        => $this->faker->sentence(),
+			'name'        => $this->faker->words( wp_rand( 2, 4 ), true ),
 			'taxonomy'    => $this->taxonomy,
 		];
 	}